backend/pyproject.toml
[project]
name = "investment-control-center-backend"
version = "0.1.0"
description = "FastAPI backend for Investment Control Center"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "gunicorn>=23.0.0",
  "sqlalchemy>=2.0.30",
  "alembic>=1.13.2",
  "psycopg[binary]>=3.2.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.4",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "celery[redis]>=5.4.0",
  "redis>=5.0.7",
  "tenacity>=8.5.0",
  "structlog>=24.2.0",
  "email-validator>=2.2.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.3.0", "httpx>=0.27.0", "ruff>=0.5.6"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
pythonpath = ["app"]
